naive area computation for triangles
