+++ autoarray/structures/triangles/abstract.py
@@ -30,6 +30,21 @@ class AbstractTriangles(ABC):
         self.vertices = vertices
 
     @property
+    def area(self) -> float:
+        """
+        The total area covered by the triangles.
+        """
+        triangles = self.triangles
+        return (
+            0.5
+            * np.abs(
+                (triangles[:, 0, 0] * (triangles[:, 1, 1] - triangles[:, 2, 1]))
+                + (triangles[:, 1, 0] * (triangles[:, 2, 1] - triangles[:, 0, 1]))
+                + (triangles[:, 2, 0] * (triangles[:, 0, 1] - triangles[:, 1, 1]))
+            ).sum()
+        )
+
+    @property
     @abstractmethod
     def numpy(self):
         pass
