+++ test_autoarray/structures/triangles/conftest.py
@@ -1,4 +1,7 @@
 from autoarray.numpy_wrapper import numpy as np
+from autoarray.structures.triangles.array import ArrayTriangles
+
+
 import pytest
 
 
@@ -17,3 +20,23 @@ def compare_with_nans():
         return np.all(equal_elements)
 
     return compare_with_nans_
+
+
+@pytest.fixture
+def triangles():
+    return ArrayTriangles(
+        indices=np.array(
+            [
+                [0, 1, 2],
+                [1, 2, 3],
+            ]
+        ),
+        vertices=np.array(
+            [
+                [0.0, 0.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+                [1.0, 1.0],
+            ]
+        ),
+    )
