+++ test_autoarray/structures/triangles/test_area.py
@@ -1,0 +1,2 @@
+def test_area(triangles):
+    assert triangles.area == 1.0
